Bab 11 - Basic authentication: Registration

+++ app/Http/Controllers/ChirpController.php
@@ -61,11 +61,14 @@ class ChirpController extends Controller
              'message.max' => 'Chirps must be 255 characters or less.',
         ]);
 
+        // Use the authenticated user
+        auth()->user()->chirps()->create($validated);
+
         // Create the chirp (no user for now - we'll add auth later)
-        \App\Models\Chirp::create([
-            'message' => $validated['message'],
-            'user_id' => null, // We'll add authentication in lesson 11
-        ]);
+        //  \App\Models\Chirp::create([
+            //  'message' => $validated['message'],
+            //  'user_id' => null, // We'll add authentication in lesson 11
+        //  ]);
 
         // Redirect back to the feed
         return redirect('/')->with('success', 'Your chirp has been posted!');
@@ -83,17 +86,22 @@ class ChirpController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(string $id)
+    public function edit(Chirp $chirp) //string $id
     {
         // We'll add authorization in lesson 11
+        // return view('chirps.edit', compact('chirp'));
+        $this->authorize('update', $chirp);
+
         return view('chirps.edit', compact('chirp'));
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, string $id)
+    public function update(Request $request, Chirp $chirp) //string $id
     {
+        $this->authorize('update', $chirp);
+    
         // Validate
         $validated = $request->validate([
             'message' => 'required|string|max:255',
@@ -108,8 +116,10 @@ class ChirpController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(Chirp $chirp) //string $id
     {
+        $this->authorize('delete', $chirp);
+
         $chirp->delete();
 
         return redirect('/')->with('success', 'Chirp deleted!');
